rollback 0.3.0 - release in new commit. Add performance improvements
