+++ .cargo/config.toml
@@ -2,4 +2,15 @@
 rustflags = [
     "-C", "link-arg=-undefined",
     "-C", "link-arg=dynamic_lookup",
+    "-C", "target-cpu=native",
+]
+
+[target.x86_64-unknown-linux-gnu]
+rustflags = [
+    "-C", "target-feature=+aes,+avx,+avx2,+sse3,+fma,+ssse3",
+]
+
+[target.aarch64-unknown-linux-gnu]
+rustflags = [
+    "-C", "target-feature=+neon,+sve,+fma",
 ]
