+++ Cargo.toml
@@ -1,6 +1,8 @@
+cargo-features = ["profile-rustflags"]
+
 [package]
-name = "polars_ols"
-version = "0.3.0"
+name = "polars_ols_worktree"
+version = "0.2.9"
 edition = "2021"
 
 [lib]
@@ -11,9 +13,9 @@ crate-type= ["cdylib"]
 pyo3 = { version = "*", features = ["extension-module", "abi3-py38"] }  # set > py38 supported version
 pyo3-polars = { version = "*", features = ["derive"] }
 serde = { version = "*", features = ["derive"] }
-polars = { version = "*", features = ["performant", "lazy", "ndarray", "dtype-struct"]}
+polars = { version = "*", features = ["performant", "lazy", "ndarray", "dtype-struct", "nightly"]}
 ndarray-rand = {version = "*"}
-faer = {version = "*", features = ["std", "rayon", "serde", "rand", "npy"]}
+faer = {version = "*", features = ["std", "rayon", "serde", "rand", "npy", "nightly"]}
 faer-ext = {version = "*", features = ["ndarray"]}
 num-traits = { version = "*" }
 approx = {version = "*" }
@@ -36,3 +38,6 @@ ndarray-linalg = {version = "*"}
 
 [profile.dev.package.faer]
 opt-level = 3
+
+[profile.release]
+lto = "thin"
